+++ www/js/library.js
@@ -76,6 +76,27 @@ angular.module('mopidy-mobile.library', [
         }
       }
     })
+    .state('tabs.library.lookup', {
+      url: '/lookup?name&uri',
+      controller: 'LookupCtrl',
+      templateUrl: 'templates/lookup.html',
+      resolve: {
+        mopidy: function(connection) {
+          return connection();
+        },
+        name: function($stateParams) {
+          return $stateParams.name;
+        },
+        uri: function($stateParams) {
+          return $stateParams.uri;
+        },
+        tracks: function($stateParams, connection) {
+          return connection(function(mopidy) {
+            return mopidy.library.lookup({uri: $stateParams.uri});
+          });
+        }
+      }
+    })
   ;
 })
 
@@ -130,11 +151,110 @@ angular.module('mopidy-mobile.library', [
   };
 })
 
-.controller('SearchCtrl', function($scope, settings, mopidy, q, results) {
-  $scope.q = q;
-  $scope.results = results;
+.controller('SearchCtrl', function($scope, $ionicPopover, settings, mopidy, q, results) {
+  function compare(a, b) {
+    if ((a.name || '') > (b.name || '')) {
+      return 1;
+    } else if ((a.name || '') < (b.name || '')) {
+      return -1;
+    } else {
+      return 0;
+    }
+  }
 
-  $scope.click = function(model) {
-    settings.click(mopidy, model.uri);
+  $scope.q = q;
+
+  switch (results.length) {
+  case 0:
+    $scope.artists = $scope.albums = $scope.tracks = [];
+    break;
+  case 1:
+    // single result - keep order
+    $scope.artists = results[0].artists;
+    $scope.albums = results[0].albums;
+    $scope.tracks = results[0].tracks;
+    break;
+  default:
+    // multiple results - merge and sort by name
+    $scope.artists = Array.prototype.concat.apply([], results.map(function(result) {
+      return result.artists || [];
+    })).sort(compare);
+    $scope.albums = Array.prototype.concat.apply([], results.map(function(result) {
+      return result.albums || [];
+    })).sort(compare);
+    $scope.tracks = Array.prototype.concat.apply([], results.map(function(result) {
+      return result.tracks || [];
+    })).sort(compare);
+  }
+
+  $ionicPopover.fromTemplateUrl('templates/popovers/library.html', {
+    scope: $scope,
+  }).then(function(popover) {
+    $scope.popover = popover;
+  });
+
+  $scope.add = function() {
+    return mopidy.tracklist.add({tracks: angular.copy($scope.tracks)});
+  };
+
+  $scope.click = function(track) {
+    return settings.click(mopidy, track.uri);
+  };
+
+  $scope.play = function() {
+    return mopidy.tracklist.add({
+      tracks: angular.copy($scope.tracks)
+    }).then(function(tlTracks) {
+      return mopidy.playback.play({tl_track: tlTracks[0]});
+    });
+  };
+
+  $scope.replace = function() {
+    return mopidy.tracklist.clear({
+      // no params
+    }).then(function() {
+      return mopidy.tracklist.add({tracks: angular.copy($scope.tracks)});
+    }).then(function(tlTracks) {
+      return mopidy.playback.play({tl_track: tlTracks[0]});
+    });
+  };
+})
+
+.controller('LookupCtrl', function($scope, $ionicPopover, settings, mopidy, name, tracks, uri) {
+  $scope.name = name;
+  $scope.tracks = tracks;
+  $scope.uri = uri;
+
+  $ionicPopover.fromTemplateUrl('templates/popovers/library.html', {
+    scope: $scope,
+  }).then(function(popover) {
+    $scope.popover = popover;
+  });
+
+
+  $scope.add = function() {
+    return mopidy.tracklist.add({tracks: angular.copy($scope.tracks)});
+  };
+
+  $scope.click = function(track) {
+    return settings.click(mopidy, track.uri);
+  };
+
+  $scope.play = function() {
+    return mopidy.tracklist.add({
+      tracks: angular.copy($scope.tracks)
+    }).then(function(tlTracks) {
+      return mopidy.playback.play({tl_track: tlTracks[0]});
+    });
+  };
+
+  $scope.replace = function() {
+    return mopidy.tracklist.clear({
+      // no params
+    }).then(function() {
+      return mopidy.tracklist.add({tracks: angular.copy($scope.tracks)});
+    }).then(function(tlTracks) {
+      return mopidy.playback.play({tl_track: tlTracks[0]});
+    });
   };
 });